1.捕获鼠标是修改图标 2.检验输入数据

+++ ProSpy/RunThread.cpp
@@ -262,10 +262,10 @@ void CRunThread::KeyInput(const OpItem* pItem)
 	}
 	if (!inputAry.empty())
 	{
-		SendInput(inputAry.size(),&inputAry[0],sizeof(INPUT));
+		SendInput((UINT)inputAry.size(),&inputAry[0],sizeof(INPUT));
 	} 
 	if (!releaseAry.empty())
 	{
-		SendInput(releaseAry.size(),&releaseAry[0],sizeof(INPUT));
+		SendInput((UINT)releaseAry.size(),&releaseAry[0],sizeof(INPUT));
 	} 
 }
